pyproject.toml
[project]
name = "tap-hotmart"
version = "1.0.0"
description = "Singer tap for the Hotmart API, built with the Meltano Singer SDK."
authors = [{ name = "Matheus", email = "matheus@datapype.com.br" }]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["singer", "meltano", "hotmart", "elt", "tap"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "singer-sdk~=0.42",
    "hotmart-python~=1.0",
]

[project.scripts]
tap-hotmart = "tap_hotmart.tap:TapHotmart.cli"

[project.urls]
Homepage = "https://github.com/im-voracity/tap-hotmart"
Repository = "https://github.com/im-voracity/tap-hotmart"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tap_hotmart"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.claude",
    "/CLAUDE.md",
    "/SPEC.md",
    "/PLAN.md",
    "/uv.lock",
    "/config.json",
]

# ---------------------------------------------------------------------------
# UV
# ---------------------------------------------------------------------------
[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "pytest-recording>=0.13",
    "vcrpy>=6.0",
    "singer-sdk[testing]~=0.42",
    "mypy>=1.10",
    "ruff>=0.5",
]

# ---------------------------------------------------------------------------
# Ruff — linting + formatting (Black-compatible, slightly adapted)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["tap_hotmart", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",  # line too long — handled by formatter
    "B008",  # do not perform function calls in default arguments
]

[tool.ruff.lint.isort]
known-first-party = ["tap_hotmart"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ---------------------------------------------------------------------------
# Mypy — strict type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = ["hotmart_python.*", "hotmart.*", "singer_sdk.*"]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "unit: fast unit tests with no external dependencies",
    "integration: tests using VCR cassettes",
    "slow: tests that take more than a few seconds",
]

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["tap_hotmart"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80